# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentPlane API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users/{user_id}/status": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Admin User Status",
        "operationId": "admin_user_status_api_v1_admin_users__user_id__status_patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/permissions": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin User Permissions",
        "operationId": "admin_user_permissions_api_v1_admin_users__user_id__permissions_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPermissionsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Admin User Permissions Replace",
        "operationId": "admin_user_permissions_replace_api_v1_admin_users__user_id__permissions_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPermissionsReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPermissionsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/agent-grants": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Admin User Agent Grants",
        "operationId": "admin_user_agent_grants_api_v1_admin_users__user_id__agent_grants_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentGrantReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPermissionsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/tool-grants": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Admin User Tool Grants",
        "operationId": "admin_user_tool_grants_api_v1_admin_users__user_id__tool_grants_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolGrantReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPermissionsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invocations": {
      "get": {
        "tags": [
          "admin-invocations"
        ],
        "summary": "Admin Invocations List",
        "operationId": "admin_invocations_list_api_v1_admin_invocations_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Application Id"
            }
          },
          {
            "name": "external_request_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "External Request Id"
            }
          },
          {
            "name": "external_user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "External User Id"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvocationDecision"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "QUEUED",
                    "RUNNING",
                    "WAITING_APPROVAL",
                    "SUCCEEDED",
                    "FAILED",
                    "CANCELLED",
                    "REJECTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvocationRead"
                  },
                  "title": "Response Admin Invocations List Api V1 Admin Invocations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invocations/{invocation_id}": {
      "get": {
        "tags": [
          "admin-invocations"
        ],
        "summary": "Admin Invocations Get",
        "operationId": "admin_invocations_get_api_v1_admin_invocations__invocation_id__get",
        "parameters": [
          {
            "name": "invocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invocation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvocationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invocations": {
      "post": {
        "tags": [
          "invocations"
        ],
        "summary": "Invocations Create",
        "operationId": "invocations_create_api_v1_invocations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvocationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/invocations/{invocation_id}": {
      "get": {
        "tags": [
          "invocations"
        ],
        "summary": "Invocations Get",
        "operationId": "invocations_get_api_v1_invocations__invocation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invocation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvocationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentGrantReplace": {
        "properties": {
          "agent_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Agent Ids"
          }
        },
        "type": "object",
        "title": "AgentGrantReplace"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvocationCreate": {
        "properties": {
          "external_request_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "External Request Id"
          },
          "external_user_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "External User Id"
          },
          "conversation_key": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Conversation Key"
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "input": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "external_request_id",
          "external_user_id",
          "conversation_key",
          "agent_id",
          "input"
        ],
        "title": "InvocationCreate"
      },
      "InvocationDecision": {
        "type": "string",
        "enum": [
          "ALLOWED",
          "DENIED"
        ],
        "title": "InvocationDecision"
      },
      "InvocationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "credential_id": {
            "type": "string",
            "format": "uuid",
            "title": "Credential Id"
          },
          "external_request_id": {
            "type": "string",
            "title": "External Request Id"
          },
          "external_user_id": {
            "type": "string",
            "title": "External User Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "conversation_key": {
            "type": "string",
            "title": "Conversation Key"
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "agent_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version Id"
          },
          "decision": {
            "$ref": "#/components/schemas/InvocationDecision"
          },
          "decision_code": {
            "type": "string",
            "title": "Decision Code"
          },
          "decision_message": {
            "type": "string",
            "title": "Decision Message"
          },
          "effective_tool_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Effective Tool Keys"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "credential_id",
          "external_request_id",
          "external_user_id",
          "user_id",
          "conversation_key",
          "agent_id",
          "agent_version_id",
          "decision",
          "decision_code",
          "decision_message",
          "effective_tool_keys",
          "session_id",
          "run_id",
          "status",
          "output",
          "error_code",
          "error_message",
          "created_at"
        ],
        "title": "InvocationRead"
      },
...
      "ToolGrantReplace": {
        "properties": {
          "tool_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Tool Keys"
          }
        },
        "type": "object",
        "title": "ToolGrantReplace"
      },
...
      "UserPermissionsRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "agent_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Agent Ids"
          },
          "tool_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tool Keys"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "agent_ids",
          "tool_keys"
        ],
        "title": "UserPermissionsRead"
      },
      "UserPermissionsReplace": {
        "properties": {
          "agent_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Agent Ids"
          },
          "tool_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Tool Keys"
          }
        },
        "type": "object",
        "title": "UserPermissionsReplace"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "login_name": {
            "type": "string",
            "title": "Login Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "login_name",
          "display_name",
          "role",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "UserRead"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "USER"
        ],
        "title": "UserRole"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACTIVE",
          "DISABLED"
        ],
        "title": "UserStatus"
      },
      "UserStatusPatch": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UserStatusPatch"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}